* moved policy code into its own module and wrote additional docs; removed dbg
