+++ src/prelude.rs
@@ -6,6 +6,7 @@ pub use crate::{
         group_api::{GroupId, GroupName},
         user_api::{DeviceId, DeviceName, UserId},
     },
+    policy::PolicyGrant,
     user::UserOps,
     DeviceContext, IronOxide, IronOxideErr,
 };
